fix(van-tabs): 修复小程序插件内使用van-tabs切换tab没有反应的问题 feat(van-tabs): incorrect style docs(van-tabs): 去除utils,version里面require循环依赖,修复插件内调用wx.canIUse抛异常问题
